+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+2.10.0 / 2019-12-12
+===================
+
+* FindByLabelText
+* LocalStorage.all
+* Browser.setMedia
+* Element getAttribute method
+
 2.9.0 / 2019-11-20
 ==================
 
@@ -44,7 +52,7 @@
 2.6.0 / 2019-08-06
 ==================
 
-* Browser.text will return newlines on </br>
+* Browser.text will return newlines on `</br>`
 * DefaultTimeout option on create browser
 * Injection scripts are now added through evaluate, so are not checked by CSP
 * Static config file removed